feat(rating): create article rating
- Implement get average article rating

[Finishes #166272109]
